autonomous_web_agent/requirements.txt
playwright>=1.40.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0